+++ CHANGES.rst
@@ -3,6 +3,7 @@ ChangeLog
 
 0.10 (TBD)
 ----------
+* [UPDATE] added byte_data argument to appdataClient.create method
 * [UPDATE] added validate_extension function to regex package
 
 0.9 (2017.04.18)